+++ src/ui/extensionsInstalled.ts
@@ -100,20 +100,25 @@ class QuartoExtensionTreeDataProvider implements vscode.TreeDataProvider<Workspa
 	}
 
 	private getWorkspaceFolderItems(): WorkspaceFolderTreeItem[] {
-		return this.workspaceFolders.map((folder) => new WorkspaceFolderTreeItem(folder.name, folder.uri.fsPath));
+		return this.workspaceFolders
+			.filter((folder) => {
+				const folderData = this.extensionsDataByFolder[folder.uri.fsPath] || {};
+				return Object.keys(folderData).length > 0;
+			})
+			.map((folder) => new WorkspaceFolderTreeItem(folder.name, folder.uri.fsPath));
 	}
 
 	private getExtensionItems(workspacePath: string): ExtensionTreeItem[] {
 		const folderData = this.extensionsDataByFolder[workspacePath] || {};
 		if (Object.keys(folderData).length === 0) {
 			return [
-				new ExtensionTreeItem(
-					"No extensions installed.",
-					vscode.TreeItemCollapsibleState.None,
-					workspacePath,
-					undefined,
-					"info"
-				),
+				// new ExtensionTreeItem(
+				// 	"No extensions installed.",
+				// 	vscode.TreeItemCollapsibleState.None,
+				// 	workspacePath,
+				// 	undefined,
+				// 	"info"
+				// ),
 			];
 		}