+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat: add a welcome view in the Quarto Wizard Explorer.
 - feat: handle empty workspace(s) in the explorer view.
+- feat: add workspace folder install context command.
 - fix: disable "blanks-around-fenced-divs" rule by default.
 - refactor(src/utils/workspace.ts): use `vscode.WorkspaceFolderPickOptions()` for workspace folder selection.
 - docs: add a note about the "blanks-around-fenced-divs" rule in the README.