+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.15.2 (2025-03-12)
 
 - feat: add a welcome view in the Quarto Wizard Explorer.
 - feat: handle empty workspace(s) in the explorer view.